feat: add less nonce size method
